Document the meaning of indices in `cut`'s `range`
The exact meaning of the indices used in the `range` of the `cut`
function/method was somewhat unclear.

During the development of https://github.com/nushell/engine-q we
slightly misused them as byte indices (obtained through `std::str::char_indices`) and thus caused panics on certain
non-ascii character containing sequences.

This PR should specify the meaning of the indices and provide a hint in
the assertion of ranges.